Tests: Renamed test_utils as test_common
